// prisma/generated/zod/inputTypeSchemas/StringNullableFilterSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { NestedStringNullableFilterSchema } from './NestedStringNullableFilterSchema';

export const StringNullableFilterSchema: z.ZodType<Prisma.StringNullableFilter> = z.object({
  equals: z.string().optional().nullable(),
  in: z.string().array().optional().nullable(),
  notIn: z.string().array().optional().nullable(),
  lt: z.string().optional(),
  lte: z.string().optional(),
  gt: z.string().optional(),
  gte: z.string().optional(),
  contains: z.string().optional(),
  startsWith: z.string().optional(),
  endsWith: z.string().optional(),
  not: z.union([ z.string(),z.lazy(() => NestedStringNullableFilterSchema) ]).optional().nullable(),
}).strict();

export default StringNullableFilterSchema;